tech-project/tests_app/requirements.txt
pytest
allure-pytest
webdriver_manager
selenium
pytest-xdist
sqlalchemy